requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

einops
hydra-core
lightning
lion_pytorch
matplotlib
numpy
omegaconf
pandas
pytorch_lightning
Requests
scikit_learn
scipy
seaborn
torch
torch_ema
transformers
wandb
kagglehub
onnx



